+++ frontend/src/Components/Buttons/BackButton.js
@@ -1,5 +1,5 @@
 import { useHistory, useLocation } from 'react-router-dom'
-import { BackIcon } from './Icons'
+import { BackIcon } from '../Icons'
 
 export default function BackButton({ onClick, buttonText }) {
   function handleOnClick() {
