+++ frontend/src/Components/Buttons/LogoutButton.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
-import { LogoutIcon } from './Icons'
+import { LogoutIcon } from '../Icons'
 import { useHistory, useLocation } from 'react-router-dom'
 
 export default function LogoutButton({ onClick }) {
