+++ frontend/src/Components/Buttons/UploadButtonFooter.js
@@ -1,7 +1,10 @@
-import { apiGetVisionLabels, apiSaveKeepit } from '../Services/apiRequests.js'
+import {
+  apiGetVisionLabels,
+  apiSaveKeepit,
+} from '../../Services/apiRequests.js'
 import { useState, useEffect } from 'react'
 import { useHistory, useLocation } from 'react-router-dom'
-import { ReactComponent as ActionButtonSvg } from '../Assets/keepit-button.svg'
+import { ReactComponent as ActionButtonSvg } from '../../Assets/keepit-button.svg'
 import styled from 'styled-components/macro'
 
 export default function UploadButtonFooter() {
