+++ frontend/src/Components/Header/Header.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
-import { ReactComponent as Logo } from '../Assets/logo.svg'
+import { ReactComponent as Logo } from '../../Assets/logo.svg'
 
 export default function Header({ onClick, buttonText }) {
   return (
