+++ frontend/src/Components/LoadingSpinner/LoadingSpinnerCloud.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
-import { CloudIcon } from './Icons'
+import { CloudIcon } from '../Icons'
 
 export default function LoadingSpinner({}) {
   return (
