+++ frontend/src/Components/Separator/ContentSeparator.js
