+++ frontend/src/Components/Tags/CustomTagForm.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
-import { PlusIcon } from './Icons'
+import { PlusIcon } from '../Icons'
 
 export default function CustomTagForm({ onSubmit }) {
   return (
