+++ frontend/src/Pages/HomePage.js
@@ -1,15 +1,15 @@
 import { useEffect, useState } from 'react'
 import styled from 'styled-components/macro'
 import { apiGetAllKeepits } from '../Services/apiRequests.js'
-import Footer from '../Components/Footer'
-import UploadButtonFooter from '../Components/UploadButtonFooter'
-import SearchButton from '../Components/SearchButton'
-import Header from '../Components/Header'
-import Taglist from '../Components/Taglist'
-import KeepitList from '../Components/KeepitList'
-import ResetFilterButton from '../Components/ResetFilterButton'
-import ContentSeparator from '../Components/ContentSeparator'
-import LogoutButton from '../Components/LogoutButton.js'
+import Header from '../Components/Header/Header'
+import Footer from '../Components/Footer/Footer'
+import UploadButtonFooter from '../Components/Buttons/UploadButtonFooter'
+import SearchButton from '../Components/Buttons/SearchButton'
+import ResetFilterButton from '../Components/Buttons/ResetFilterButton'
+import LogoutButton from '../Components/Buttons/LogoutButton.js'
+import KeepitList from '../Components/KeepitList/KeepitList'
+import Taglist from '../Components/Tags/Taglist'
+import ContentSeparator from '../Components/Separator/ContentSeparator'
 import { FilterIcon } from '../Components/Icons'
 
 export default function HomePage() {
