+++ frontend/src/Pages/NewKeepitPage.js
@@ -1,18 +1,18 @@
 import { useState, useEffect } from 'react'
 import { useHistory } from 'react-router-dom'
-import { StarIcon, TagIcon, DoneIcon, DeleteIcon } from '../Components/Icons'
 import styled from 'styled-components/macro'
 
-import TaglistNewKeepit from '../Components/TaglistNewKeepit'
-import CustomTagForm from '../Components/CustomTagForm'
-import Footer from '../Components/Footer'
-import BackButton from '../Components/BackButton'
-import SearchButton from '../Components/SearchButton'
-import SaveButtonFooter from '../Components/SaveButtonFooter'
-import Header from '../Components/Header'
-import StarRatingForm from '../Components/StarRatingForm'
-import Overlay from '../Components/Overlay'
-import ContentSeparator from '../Components/ContentSeparator'
+import Header from '../Components/Header/Header'
+import Footer from '../Components/Footer/Footer'
+import TaglistNewKeepit from '../Components/Tags/TaglistNewKeepit'
+import CustomTagForm from '../Components/Tags/CustomTagForm'
+import BackButton from '../Components/Buttons/BackButton'
+import SearchButton from '../Components/Buttons/SearchButton'
+import SaveButtonFooter from '../Components/Buttons/SaveButtonFooter'
+import StarRatingForm from '../Components/StarRating/StarRatingForm'
+import Overlay from '../Components/Overlay/Overlay'
+import ContentSeparator from '../Components/Separator/ContentSeparator'
+import { StarIcon, TagIcon, DoneIcon, DeleteIcon } from '../Components/Icons'
 
 import setBase64 from '../Util/setBase64'
